requirements.txt
setuptools~=65.0.0
gym~=0.21.0
numpy~=1.24.4
matplotlib~=3.2.2
keyboard~=0.13.5
gym-retro==0.8.0
stable-baselines3==1.7.0
tensorboard==2.12.1